# openapi.yaml
openapi: 3.0.3
info:
  title: CatWAF API
  version: 1.0.1
  description: |
    CatWAF's control-plane API — everything the dashboard itself calls.
    There's no separate "public API" here; this is the real thing the
    frontend uses, documented so you can script against it or build your
    own tooling on top.

    **Scope of this document:** CatWAF has ~167 routes across 19 files.
    This spec covers the core groups in real depth — auth and two-factor
    login, WAF engine settings, custom rules, IP/geo lists, the security
    score, honeypots, sensitive-file protection, operational modes
    (panic/maintenance/config signing), and the settings, protection and
    operations layers — rather than a shallow stub for every endpoint.
    Routes not yet documented here still exist and work; see
    `backend/routes/` for the full list. Expanding this file's coverage is
    tracked in `ROADMAP.md`.

    **The settings layer is self-describing.** Rather than enumerating ~290
    fields here, `GET /api/settings/schema` returns the live schema the
    backend validates against and the dashboard builds its controls from —
    field names, types, defaults, permitted values and help text. That is
    the authoritative reference; `docs/settings.md` is generated from it.

    **One thing worth knowing:** the dashboard stats/traffic endpoints
    (`/api/stats`, `/api/traffic/*`) return real data — sourced from
    Coraza's own audit log via `services/requestLog.js`, not a demo feed —
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Obtained from POST /api/auth/login. Unauthenticated requests are
        still accepted for GET routes (treated as the built-in `guest`/
        viewer role) — the token only matters for anything that mutates
        state, which requires the `admin` role.
  responses:
    Unauthorized:
      description: Read-only role — log in as admin to modify this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: Login rate limit exceeded (10 attempts / 5 minutes per IP+username).
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until the limit resets }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        detail: { type: string, example: 'Read-only role. Log in as admin to modify.' }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, format: password, example: admin }
    LoginResponse:
      type: object
      properties:
        token: { type: string, description: 'JWT — send as `Authorization: Bearer <token>` on subsequent requests' }
        user:
          type: object
          properties:
            id: { type: string }
            username: { type: string }
            role: { type: string, enum: [admin, viewer] }
...
    AppliedRoute:
      type: object
      description: One route in CatWAF's generated region of the live Caddyfile.
      properties:
        name: { type: string }
        runtime: { type: string, nullable: true, description: 'Detected runtime, e.g. "PHP 8.3"' }
        containerName: { type: string, nullable: true }
        listenPort: { type: integer, description: Port CatWAF listens on for this application }
        upstream: { type: string, nullable: true, description: Address CatWAF forwards to once a request passes the WAF }
    DiscoveredApp:
      type: object
      description: >
        The safe projection of a discovered container. Raw environment and
        labels are deliberately absent — see backend/services/discovery/summary.js.
      properties:
        name: { type: string, description: Compose service name where there is one, otherwise the container name }
        containerName: { type: string }
        image: { type: string }
        running: { type: boolean }
        isWeb: { type: boolean, description: Whether this serves HTTP and is therefore routable }
        webServer: { type: string, nullable: true }
        webServerConfidence: { type: number, nullable: true }
        runtime: { type: string, nullable: true }
        framework: { type: string, nullable: true }
        port: { type: integer, nullable: true }
        published: { type: boolean, description: Published to the host — such a port bypasses CatWAF while it stays published }
        reachability: { type: string, nullable: true }
        networks: { type: array, items: { type: string } }
        phpConfidence: { type: number, nullable: true }
        fastcgiBackend: { type: string, nullable: true }
        fpmBackendFor: { type: array, nullable: true, items: { type: string } }
        incomplete: { type: boolean }
        evidence:
          type: object
          description: Why CatWAF classified it this way. Files, processes, ports and config directives — never secret values.
          properties:
            web: { type: array, items: { type: string } }
            php: { type: array, items: { type: string } }
            runtime: { type: array, items: { type: string } }
            server: { type: array, items: { type: string } }
    VerificationResult:
      type: object
      properties:
        name: { type: string }
        protected: { type: boolean, description: True only when a benign request was served AND a CRS payload was refused }
        reason: { type: string, nullable: true, description: Why the route is not protected }
        benignStatus: { type: integer, nullable: true }
        attackStatus: { type: integer, nullable: true }
        upstreamUnreachable: { type: boolean, description: 'CatWAF is listening but the application behind it is down — the site is broken, not merely unprotected' }
        wafBypassed: { type: boolean, description: 'A CRS payload reached the application, so Coraza is not inspecting this route' }
    Verification:
      type: object
      properties:
        results: { type: array, items: { $ref: '#/components/schemas/VerificationResult' } }
        protectedCount: { type: integer }
        allProtected: { type: boolean }
    ProtectResult:
      type: object
      properties:
        ok: { type: boolean }
        stage:
          type: string
          description: How far the pipeline got.
          enum: [discovery, plan, dry-run, apply, applied, unverified, protected]
        dry_run: { type: boolean }
        error: { type: string, nullable: true }
        message: { type: string, nullable: true }
        containers: { type: array, items: { $ref: '#/components/schemas/DiscoveredApp' } }
        web_apps: { type: array, items: { $ref: '#/components/schemas/DiscoveredApp' } }
        routes: { type: array, items: { $ref: '#/components/schemas/AppliedRoute' } }
        skipped:
          type: array
          description: Applications CatWAF found but could not route to, with diagnostics.
          items:
            type: object
            properties:
              name: { type: string }
              reason: { type: string }
              diagnostics: { type: array, items: { type: string } }
        backup: { type: string, nullable: true, description: Path the previous Caddyfile was copied to }
        reloaded: { type: boolean, nullable: true }
        reload_error: { type: string, nullable: true }
        verification: { allOf: [{ $ref: '#/components/schemas/Verification' }], nullable: true }
        protected:
          type: boolean
          description: >
            Derived solely from `verification.allProtected`. A generated,
            applied and cleanly reloaded configuration does not set this.
...
paths:
...
  /api/apps:
    get:
      tags: [Applications]
      summary: Routes CatWAF has applied, read back from the live Caddyfile
      description: >
        Side-effect free, safe to poll. Reports what is *configured*. `hasRegion`
        distinguishes "no generated region in the Caddyfile" from "CatWAF could
        not read the Caddyfile at all", which `ok: false` reports instead.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, description: false when the Caddyfile could not be read }
                  error: { type: string, nullable: true }
                  caddyfile: { type: string, description: Path that was read }
                  hasRegion: { type: boolean, description: Whether CatWAF's generated region exists }
                  caddy_running: { type: boolean }
                  busy: { type: string, nullable: true, description: Name of a pipeline run in progress, if any }
                  routes:
                    type: array
                    items: { $ref: '#/components/schemas/AppliedRoute' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/apps/discover:
    post:
      tags: [Applications]
      summary: Inspect this host for web applications (OBSERVE -> CORRELATE -> CLASSIFY)
      description: >
        Writes nothing. Answers 200 with `ok: false` and a reason when the host
        cannot be inspected (no Docker daemon, no permission) — that is an
        outcome, not a transport error. Container environment variables and
        labels are never returned: they routinely hold database passwords.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                skip_http_probe:
                  type: boolean
                  default: false
                  description: Skip the per-port HTTP probe. Faster, but "a port is open" is then not confirmed to be "a port serving HTTP".
      responses:
        '200':
          description: Discovery ran, or explained why it could not
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  code: { type: string, description: Machine-readable reason when ok is false }
                  detail: { type: string, description: Human-readable reason when ok is false }
                  containers: { type: array, items: { $ref: '#/components/schemas/DiscoveredApp' } }
                  web_apps: { type: array, items: { $ref: '#/components/schemas/DiscoveredApp' } }
                  errors:
                    type: array
                    description: Containers that could not be inspected; the rest of the scan still completed.
                    items:
                      type: object
                      properties: { name: { type: string }, error: { type: string } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Another pipeline run is already in progress. }
  /api/apps/preview:
    post:
      tags: [Applications]
      summary: The routes a protect run would create — nothing is written
      description: >
        A full dry run: discovery, upstream resolution and route planning, with
        no file written, no reload, and no port assignment persisted. `protected`
        is always false here by construction.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/ProtectResult' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Another pipeline run is already in progress. }
  /api/apps/protect:
    post:
      tags: [Applications]
      summary: Apply the generated configuration, reload, then verify with real traffic
      description: >
        Admin only, and audited. Rewrites CatWAF's generated region, backs up the
        previous Caddyfile, reloads, then proves the result with a real benign
        request and a real inert CRS payload against each route.


        Unlike `catwaf auto`, this never starts the proxy: that is a privileged
        host action. If Caddy is not running the response reports the routes as
        applied but not live, rather than trying to start services.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                skip_verify:
                  type: boolean
                  default: false
                  description: Apply without verifying. `protected` is then false, because nothing was proven.
      responses:
        '200':
          description: The pipeline ran; read `stage`, `verification` and `protected` for the outcome.
          content: { application/json: { schema: { $ref: '#/components/schemas/ProtectResult' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Read-only role. }
        '409': { description: Another pipeline run is already in progress. }
  /api/apps/verify:
    post:
      tags: [Applications]
      summary: Prove protection with real requests
      description: >
        Sends two real requests to each applied route: a benign one that must be
        served, and the standard inert OWASP CRS SQL-injection signature that
        must be refused. A route is `protected` only if both hold. This is the
        only endpoint whose answer may be presented to a user as "protected".
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  detail: { type: string, description: Present when there is nothing applied to verify }
                  routes: { type: array, items: { $ref: '#/components/schemas/AppliedRoute' } }
                  verification: { $ref: '#/components/schemas/Verification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Another pipeline run is already in progress. }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      description: Rate limited to 10 attempts per 5 minutes, keyed by IP + username.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...